Build fixes for MacOS X
Under MacOSX, static compilation is not possible and flex and bison have
to be taken from homebrew. We now use the most plausible homebrew paths
by default under MacOSX.

For comfort, these paths can also be overridden.
These make it easier to override the default versions of these
programs, which is particularly helpful on Mac, because the versions
there are so old.